requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
flake8==7.1.0
idna==3.7
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
